requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
attrs==20.3.0
    # via -r requirements/requirements.in
billiard==3.6.4.0
    # via celery
celery-progress==0.1.0
    # via -r requirements/requirements.in
celery[redis]==5.0.5
    # via -r requirements/requirements.in
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.1.6
    # via celery
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
coreapi==2.3.3
    # via -r requirements/requirements.in
coreschema==0.0.4
    # via coreapi
cryptography==3.4.7
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
dj-stripe==2.4.4
    # via -r requirements/requirements.in
django-allauth==0.44.0
    # via -r requirements/requirements.in
django-anymail[mailgun]==8.2
    # via -r requirements/requirements.in
django==3.2.3
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   djangorestframework
    #   jsonfield
djangorestframework==3.12.4
    # via -r requirements/requirements.in
idna==2.10
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==2.11.3
    # via coreschema
jsonfield==3.1.0
    # via dj-stripe
kombu==5.0.2
    # via celery
markupsafe==1.1.1
    # via jinja2
oauthlib==3.1.0
    # via requests-oauthlib
prompt-toolkit==3.0.18
    # via click-repl
psycopg2-binary==2.8.6
    # via -r requirements/requirements.in
pycparser==2.20
    # via cffi

    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pytz==2021.1
    # via
    #   celery
    #   django
redis==3.5.3
    # via celery
requests-oauthlib==1.3.0
    # via django-allauth
requests==2.25.1
    # via
    #   coreapi
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   stripe
six==1.15.0
    # via click-repl
sqlparse==0.4.1
    # via django
stripe==2.56.0
    # via
    #   -r requirements/requirements.in
    #   dj-stripe
uritemplate==3.0.1
    # via coreapi
urllib3==1.26.4
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
wcwidth==0.2.5
    # via prompt-toolkit
PyJWT==2.1.0
django-cors-headers==3.7.0
channels==3.0.4
channels-redis==3.3.0
sentry-sdk==1.3.0